+++ src/main/java/dailymissionproject/demo/common/config/CustomOAuth2UserService.java
@@ -1,33 +1,15 @@
-package dailymissionproject.demo.config;
+package dailymissionproject.demo.common.config;
 
-import dailymissionproject.demo.config.auth.dto.OAuthAttributes;
-import dailymissionproject.demo.config.auth.dto.SessionUser;
-import dailymissionproject.demo.entity.User;
-import dailymissionproject.demo.repository.UserRepository;
-import jakarta.servlet.http.HttpSession;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserService;
-import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
-import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.stereotype.Service;
+//import dailymissionproject.demo.config.auth.dto.OAuthAttributes;
 
-import java.util.Collections;
-
+/*
 @Slf4j
 @Service
+@RequiredArgsConstructor
 public class CustomOAuth2UserService implements OAuth2UserService<OAuth2UserRequest, OAuth2User> {
 
     private final UserRepository userRepository;
-    private final HttpSession httpSession;
 
-    public CustomOAuth2UserService(UserRepository userRepository, HttpSession httpSession){
-        this.userRepository = userRepository;
-        this.httpSession = httpSession;
-    }
     @Override
     public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException{
         OAuth2UserService<OAuth2UserRequest, OAuth2User> delegate = new DefaultOAuth2UserService();
@@ -39,10 +21,10 @@ public class CustomOAuth2UserService implements OAuth2UserService<OAuth2UserRequ
 
         OAuthAttributes attributes = OAuthAttributes.of(registrationId, userNameAttributeName, oAuth2User.getAttributes());
 
-        User user = saveOrUpdate(attributes);
-        httpSession.setAttribute("user", new SessionUser(user));
+        Map<String, Object> userAttribute = attributes.convertToMap();
 
-        //System.out.println(attributes.getAttributes());
+        User user = saveOrUpdate(attributes);
+
         return new DefaultOAuth2User(
                 Collections.singleton(new SimpleGrantedAuthority(user.getRoleKey())),
                 attributes.getAttributes(),
@@ -57,3 +39,6 @@ public class CustomOAuth2UserService implements OAuth2UserService<OAuth2UserRequ
         return userRepository.save(user);
     }
 }
+
+
+ */
