+++ src/main/java/dailymissionproject/demo/common/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package dailymissionproject.demo.config;
+package dailymissionproject.demo.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -10,6 +10,7 @@ import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.servlet.util.matcher.MvcRequestMatcher;
 import org.springframework.web.servlet.handler.HandlerMappingIntrospector;
 
+/*
 @Configuration
 @EnableWebSecurity
 public class SecurityConfig {
@@ -38,3 +39,5 @@ public class SecurityConfig {
        return http.build();
     }
 }
+
+ */
