--- src/main/java/dailymissionproject/demo/controller/MissionController.java
@@ -1,9 +1,0 @@
-package dailymissionproject.demo.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-
-@Controller
-@RequiredArgsConstructor
-public class MissionController {
-}
