+++ src/main/java/dailymissionproject/demo/domain/mission/repository/Mission.java
@@ -1,10 +1,12 @@
-package dailymissionproject.demo.entity;
+package dailymissionproject.demo.domain.mission.repository;
 
+import dailymissionproject.demo.domain.participant.repository.Participant;
+import dailymissionproject.demo.domain.post.repository.Post;
+import dailymissionproject.demo.domain.user.repository.User;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.cglib.core.Local;
 import org.springframework.format.annotation.DateTimeFormat;
 import java.time.LocalDate;
 import java.util.ArrayList;
