+++ src/main/java/dailymissionproject/demo/domain/participant/repository/Participant.java
@@ -1,5 +1,7 @@
-package dailymissionproject.demo.entity;
+package dailymissionproject.demo.domain.participant.repository;
 
+import dailymissionproject.demo.domain.mission.repository.Mission;
+import dailymissionproject.demo.domain.user.repository.User;
 import jakarta.persistence.*;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
@@ -7,10 +9,10 @@ import org.springframework.data.annotation.LastModifiedDate;
 import java.time.LocalDateTime;
 
 @Entity
-public class Post {
+public class Participant {
 
     @Id @GeneratedValue
-    @Column(name = "post_id")
+    @Column(name = "participant_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
@@ -21,12 +23,7 @@ public class Post {
     @JoinColumn(name = "user_id")
     private User user;
 
-    private String title;
-    private String content;
-    @Column(name = "image_url")
-    private String imageUrl;
-
-    private boolean deleted;
+    private boolean banned;
 
     @CreatedDate
     @Column(name = "created_date")
